priv: fix memory leak in jrpc_process()
We have to free the main received object always.
This will automatically free the params object too, so user doen't have to.
This introduces a corner case, were, if user wants to pass params further,
there will be a double free, which is not what we want. So user has to
copy the params object for further use (and free it, if not passed to
jrpc_send() in method handler). Reflect that in server example.

Signed-off-by: Roman Yeryomin <roman@advem.lv>